Fix github api

+++ .ci/after_failure.sh
@@ -4,7 +4,9 @@ if [ "$TRAVIS_PULL_REQUEST" != "false" ] ; then
     if [ "$LINT" = "vimlint-errors" ] ; then
         VIMLINT_LOG=`cat build_log`
         if [ -n "$VIMLINT_LOG" ]; then
-            rm -rf build/GitHub.vim
+            if [[ -d build/GitHub.vim ]]; then
+                rm -rf build/GitHub.vim
+            fi
             git clone https://github.com/wsdjeg/GitHub.vim.git build/GitHub.vim
             vim -u .ci/common/github_commenter.vim
         fi
